requirements.txt
Flask==2.1.3
gevent==21.8.0
keras==2.9.0
numpy==1.21.5
Pillow==9.3.0
Werkzeug==2.2.2
tensorflow_cpu==2.9.0
gunicorn-20.1.0